parse field config with serde_multikey
