+++ quickwit/quickwit-doc-mapper/Cargo.toml
@@ -31,6 +31,7 @@ typetag = { workspace = true }
 utoipa = { workspace = true }
 
 quickwit-datetime = { workspace = true }
+quickwit-macros = { workspace = true }
 quickwit-query = { workspace = true }
 
 [dev-dependencies]
