+++ quickwit/quickwit-macros/Cargo.toml
@@ -14,5 +14,7 @@ proc-macro = true
 
 [dependencies]
 proc-macro2 = { workspace = true }
+quote = { workspace = true }
+syn = { workspace = true }
 
 quickwit-macros-impl = { workspace = true }
