+++ quickwit/quickwit-metastore/test-data/index-metadata/v0.5.expected.json
@@ -11,7 +11,9 @@
         "expand_dots": true,
         "fast": true,
         "indexed": true,
-        "stored": true
+        "record": "basic",
+        "stored": true,
+        "tokenizer": "raw"
       },
       "field_mappings": [
         {
@@ -39,6 +41,7 @@
           "fieldnorms": false,
           "indexed": true,
           "name": "log_level",
+          "record": "basic",
           "stored": true,
           "tokenizer": "raw",
           "type": "text"
